requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.6.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pymongo==4.8.0
PySocks==1.7.1
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
tldextract==5.1.2
urllib3==2.2.2
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.1
